Run the full SciPy test suite via scipy-tests
Rebase onto main after scipy-tests landed. Keep scipy runtime-only,
skip wasm-unsupported tests in conftest.py, and call scipy.test()
from the scipy-tests package.
